# openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Data Capture Service API",
        "description": "Manage a questionnaire lifecycle",
        "version": "2.0.0",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "API Support",
            "email": "api@cica.gov.uk"
        }
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
        "/questionnaires": {
            "post": {
                "tags": ["Questionnaires"],
                "summary": "Create a new questionnaire from a specified template",
                "description": "Create a questionnaire",
                "operationId": "createQuestionnaires",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/apiVersion"
                    }
                ],
                "requestBody": {
                    "description": "Questionnaire template id",
                    "required": true,
                    "content": {
                        "application/vnd.api+json": {
                            "schema": {
                                "$ref": "./json-schemas/api/_questionnaires/post/req/201.json"
                            },
                            "example": {
                                "data": {
                                    "type": "questionnaires",
                                    "attributes": {
                                        "templateName": "apply-for-compensation",
                                        "owner": {
                                            "id": "urn:uuid:f81d4fae-7dec-11d0-a765-00a0c91e6bf6",
                                            "isAuthenticated": true
                                        },
                                        "origin": {
                                            "channel": "web"
                                        },
                                        "external": {
                                            "id": "urn:uuid:f81d4fae-7dec-11d0-a765-123456781234"
                                        },
                                        "templateVersion": "1.0.0",
                                        "userData": {
                                            "personalisation": {
                                                "first-name": "Test",
                                                "surname": "Test case"
                                            },
                                            "caseReference": "123456"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "./json-schemas/api/_questionnaires/post/res/201.json"
                                },
                                "example": {
                                    "data": {
                                        "type": "questionnaires",
                                        "id": "todo-p-some-id",
                                        "attributes": {
                                            "sections": {},
                                            "routes": {}
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/questionnaires/{questionnaireId}/progress-entries": {
            "get": {
                "tags": ["ProgressEntries"],
                "summary": "Get progress entries for a questionnaire.",
                "description": "Gets all progress entries for a questionnaire. The results can be filtered.",
                "operationId": "getProgressEntries",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/questionnaireId"
                    },
                    {
                        "$ref": "#/components/parameters/filter"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/onBehalfOf"
                    },
                    {
                        "$ref": "#/components/parameters/apiVersion"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "json-schemas/api/_questionnaires_{questionnaireId}_progress-entries/get/res/200.json"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "type": "progress-entries",
                                            "id": "08dbd0d2-6cee-49e5-a55f-35b31ac4aa9e",
                                            "attributes": {
                                                "sectionId": "p-applicant-declaration"
                                            },
                                            "relationships": {
                                                "section": {
                                                    "data": {
                                                        "type": "sections",
                                                        "id": "p-applicant-declaration"
                                                    }
                                                }
                                            }
                                        }
                                    ],
                                    "links": {
                                        "prev": "questionnaires/{questionnaireId}/progress-entries?page[before]=p-applicant-declaration"
                                    },
                                    "included": [
                                        {
                                            "type": "sections",
                                            "id": "p-applicant-declaration",
                                            "attributes": {
                                                "$schema": "http://json-schema.org/draft-07/schema#",
                                                "type": "object",
                                                "title": "Declaration",
                                                "additionalProperties": false,
                                                "properties": {
                                                    "applicant-declaration": {
                                                        "description": "\n                <p class=\"g-body\">By continuing you confirm that the information you will give is true as far as you know.</p>\n                {{ govukWarningText({\n                    text: \"If you deliberately give false or misleading information, you may get less compensation or be prosecuted.\",\n                    iconFallbackText: \"Warning\"\n                }) }}\n            "
                                                    }
                                                }
                                            },
                                            "relationships": {
                                                "answer": {
                                                    "data": {
                                                        "type": "answers",
                                                        "id": "p-applicant-declaration"
                                                    }
                                                }
                                            }
                                        },
                                        {
                                            "type": "answers",
                                            "id": "p-applicant-declaration",
                                            "attributes": {}
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/questionnaires/{questionnaireId}/sections/{sectionId}/answers": {
            "post": {
                "tags": ["Answers"],
                "summary": "Create answer(s) for a specified section",
                "description": "Adds the supplied answer(s) to a section",
                "operationId": "createAnswers",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/questionnaireId"
                    },
                    {
                        "$ref": "#/components/parameters/sectionId"
                    },
                    {
                        "$ref": "#/components/parameters/onBehalfOf"
                    },
                    {
                        "$ref": "#/components/parameters/apiVersion"
                    }
                ],
                "requestBody": {
                    "description": "Application reference number",
                    "required": true,
                    "content": {
                        "application/vnd.api+json": {
                            "schema": {
                                "$ref": "./json-schemas/api/_questionnaires_{questionnaireId}_sections_{sectionId}_answers/post/req/201.json"
                            },
                            "example": {
                                "data": {
                                    "type": "answers",
                                    "attributes": {
                                        "case-reference": "11\\111111"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "./json-schemas/api/_questionnaires_{questionnaireId}_sections_{sectionId}_answers/post/res/201.json"
                                },
                                "example": {
                                    "data": {
                                        "type": "answers",
                                        "id": "p-some-id",
                                        "attributes": {
                                            "case-reference": "11\\111111"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "get": {
                "tags": ["Answers"],
                "summary": "Get answer(s) for a specified section",
                "description": "Gets the answer(s) to a supplied section",
                "operationId": "getAnswer",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/questionnaireId"
                    },
                    {
                        "$ref": "#/components/parameters/sectionId"
                    },
                    {
                        "$ref": "#/components/parameters/onBehalfOf"
                    },
                    {
                        "$ref": "#/components/parameters/apiVersion"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "json-schemas/api/_questionnaires_{questionnaireId}_sections_{sectionId}_answers/get/res/200.json"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "type": "answers",
                                            "id": "p-applicant-british-citizen-or-eu-national",
                                            "attributes": {
                                                "q-applicant-british-citizen-or-eu-national": true
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
...
        "/questionnaires/{questionnaireId}/template-metadata": {
            "get": {
                "tags": ["templateMetadata"],
                "summary": "Get template metadata for a questionnaire.",
                "description": "Gets all template metadata for a questionnaire. Returns undefined if none found.",
                "operationId": "getTemplateMetadata",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/questionnaireId"
                    },
                    {
                        "$ref": "#/components/parameters/onBehalfOf"
                    },
                    {
                        "$ref": "#/components/parameters/apiVersion"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "json-schemas/api/_questionnaires_{questionnaireId}_metadata/get/res/200.json"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "type": "personalisationData",
                                            "id": "c90319ef-c193-4c9e-9a95-22121848a723",
                                            "attributes": {
                                                "personalisation": {
                                                    "questionnaire-id": "c90319ef-c193-4c9e-9a95-22121848a723",
                                                    "first-name": "test",
                                                    "last-name": "testcase"
                                                },
                                                "summaryBlocks": {
                                                    "read": {
                                                        "condition": "unopened",
                                                        "link": "<a>click here</a>"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "parameters": {
            "apiVersion": {
                "name": "Dcs-Api-Version",
                "in": "header",
                "required": true,
                "description": "The date based version of the requested API",
                "schema": {
                    "$ref": "./json-schemas/models/definitions/api-version.json"
                },
                "example": "2023-05-17"
            },
            "questionnaireId": {
                "name": "questionnaireId",
                "in": "path",
                "required": true,
                "description": "The id of the specific questionnaire instance. Format UUIDv4.",
                "schema": {
                    "$ref": "./json-schemas/models/definitions/uuid-v4.json"
                },
                "example": "285cb104-0c15-4a9c-9840-cb1007f098fb"
            },
            "filter": {
                "in": "query",
                "name": "filter",
                "description": "Filter progress entries",
                "schema": {
                    "$ref": "./json-schemas/api/parameters/filter.json"
                },
                "style": "deepObject",
                "explode": true,
                "example": {
                    "position": "current"
                }
            },
            "page": {
                "in": "query",
                "name": "page",
                "description": "Paginate progress entries",
                "schema": {
                    "$ref": "./json-schemas/api/parameters/page.json"
                },
                "style": "deepObject",
                "explode": true,
                "example": {
                    "before": "p-applicant-declaration"
                }
            },
            "onBehalfOf": {
                "in": "header",
                "name": "On-Behalf-Of",
                "required": true,
                "description": "Specify the owner of the resource.",
                "schema": {
                    "$ref": "./json-schemas/models/definitions/owner-id.json"
                },
                "example": "urn:uuid:f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
            },
            "sectionId": {
                "name": "sectionId",
                "in": "path",
                "required": true,
                "description": "The id of the specific questionnaire section.",
                "schema": {
                    "$ref": "./json-schemas/models/definitions/section-id.json"
                },
                "example": "system"
            }
        },
        "responses": {
            "BadRequest": {
                "description": "There is an issue with the request",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/400.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 400,
                                    "title": "400 Bad Request",
                                    "detail": "Request JSON is malformed"
                                }
                            ]
                        }
                    }
                }
            },
            "UnauthorizedError": {
                "description": "Access token is missing or invalid",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/401.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 401,
                                    "detail": "No authorization token was found"
                                }
                            ]
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "The JWT doesn't permit access to this endpoint",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/403.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 403,
                                    "title": "403 Forbidden",
                                    "detail": "Insufficient scope"
                                }
                            ]
                        }
                    }
                }
            },
            "NotFound": {
                "description": "The specified resource was not found",
                "content": {
                    "application/vnd.api+json": {
                        "schema": {
                            "$ref": "./json-schemas/api/responses/404.json"
                        },
                        "example": {
                            "errors": [
                                {
                                    "status": 404,
                                    "title": "404 Not Found",
                                    "detail": "Resource /api/v1/questionnaires/2d7caf89-2c72-469f-b19d-17f2a22270b6/sections/answers does not exist"
                                }
                            ]
                        }
                    }
                }
            },
...
        }
    }
}